# openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponseDto": {
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "format": "uuid",
            "type": "string"
          },
          "fieldErrors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "retryAfterSeconds": {
            "minimum": 1,
            "type": "number"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "correlationId"
        ],
        "type": "object"
      },
...
      "ClosePriceRequestDto": {
        "properties": {
          "effectiveTo": {
            "format": "date-time",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "effectiveTo",
          "reason"
        ],
        "type": "object"
      },
...
      "CreateOverrideRequestDto": {
        "properties": {
          "amountMinor": {
            "pattern": "^(?:0|[1-9]\\d*)$",
            "type": "string"
          },
          "effectiveFrom": {
            "format": "date-time",
            "type": "string"
          },
          "effectiveTo": {
            "format": "date-time",
            "type": "string"
          },
          "productId": {
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "amountMinor",
          "effectiveFrom",
          "productId",
          "unitId",
          "reason"
        ],
        "type": "object"
      },
...
      "OverrideListResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OverrideResponseDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "OverrideResponseDto": {
        "properties": {
          "amountMinor": {
            "pattern": "^(?:0|[1-9]\\d*)$",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "effectiveFrom": {
            "format": "date-time",
            "type": "string"
          },
          "effectiveTo": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "householdId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "productId": {
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "vendorId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "vendorId",
          "productId",
          "unitId",
          "amountMinor",
          "effectiveFrom",
          "effectiveTo",
          "createdAt",
          "updatedAt",
          "householdId",
          "reason",
          "currency"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "opaqueBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "refreshCookie": {
        "in": "cookie",
        "name": "milktrack_refresh",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "MilkTrack API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/vendors/{vendorId}/households/{householdId}/price-overrides": {
      "get": {
        "operationId": "PriceOverrideController_list",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "productId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "unitId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideListResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "opaqueBearer": []
          }
        ],
        "tags": [
          "Customer price overrides"
        ]
      },
      "post": {
        "operationId": "PriceOverrideController_create",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOverrideRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "opaqueBearer": []
          }
        ],
        "tags": [
          "Customer price overrides"
        ]
      }
    },
    "/v1/vendors/{vendorId}/households/{householdId}/price-overrides/{overrideId}": {
      "get": {
        "operationId": "PriceOverrideController_get",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "overrideId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "opaqueBearer": []
          }
        ],
        "tags": [
          "Customer price overrides"
        ]
      }
    },
    "/v1/vendors/{vendorId}/households/{householdId}/price-overrides/{overrideId}/close": {
      "post": {
        "operationId": "PriceOverrideController_close",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "overrideId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosePriceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "opaqueBearer": []
          }
        ],
        "tags": [
          "Customer price overrides"
        ]
      }
    },
...
  },
  "servers": [],
...
}